Fix: ellipsetool should draw ellipses when in corner to corner mode.
